# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Peace Power API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserPublic {
  id: number;
  name: string;
  email: string;
  isAdmin: boolean;
  createdAt: string;
}

export interface SubmitScanBody {
  heartRate: number;
  rmssd: number;
  sdnn: number;
  coherenceScore: number;
  coherenceLevel: string;
  quality: string;
}

export interface ScanRecord {
  id: number;
  userId: number;
  heartRate: number;
  rmssd: number;
  sdnn: number;
  coherenceScore: number;
  coherenceLevel: string;
  quality: string;
  scannedAt: string;
  createdAt: string;
}

export interface MemberWithStats {
  id: number;
  name: string;
  email: string;
  isAdmin: boolean;
  createdAt: string;
  scanCount: number;
  lastScan: ScanRecord | null;
  /** @nullable */
  avgCoherenceScore: number | null;
  /** @nullable */
  avgHeartRate: number | null;
}

export interface AdminStats {
  totalMembers: number;
  totalScans: number;
  /** @nullable */
  avgCoherenceScore: number | null;
  /** @nullable */
  avgHeartRate: number | null;
  highCoherenceCount: number;
  mediumCoherenceCount: number;
  lowCoherenceCount: number;
}

export interface SetAdminRoleBody {
  isAdmin: boolean;
}