# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Peace Power API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const RegisterBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  isAdmin: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  isAdmin: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Submit a scan result
 */
export const SubmitScanBody = zod.object({
  heartRate: zod.number(),
  rmssd: zod.number(),
  sdnn: zod.number(),
  coherenceScore: zod.number(),
  coherenceLevel: zod.string(),
  quality: zod.string(),
});

/**
 * @summary Get my scan history
 */
export const GetMyScansResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  heartRate: zod.number(),
  rmssd: zod.number(),
  sdnn: zod.number(),
  coherenceScore: zod.number(),
  coherenceLevel: zod.string(),
  quality: zod.string(),
  scannedAt: zod.date(),
  createdAt: zod.date(),
});
export const GetMyScansResponse = zod.array(GetMyScansResponseItem);

/**
 * @summary List all members with stats (admin only)
 */
export const GetAdminMembersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  isAdmin: zod.boolean(),
  createdAt: zod.date(),
  scanCount: zod.number(),
  lastScan: zod.union([
    zod.object({
      id: zod.number(),
      userId: zod.number(),
      heartRate: zod.number(),
      rmssd: zod.number(),
      sdnn: zod.number(),
      coherenceScore: zod.number(),
      coherenceLevel: zod.string(),
      quality: zod.string(),
      scannedAt: zod.date(),
      createdAt: zod.date(),
    }),
    zod.null(),
  ]),
  avgCoherenceScore: zod.number().nullable(),
  avgHeartRate: zod.number().nullable(),
});
export const GetAdminMembersResponse = zod.array(GetAdminMembersResponseItem);

/**
 * @summary Get all scans for a member (admin only)
 */
export const GetAdminMemberScansParams = zod.object({
  userId: zod.coerce.number(),
});

export const GetAdminMemberScansResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  heartRate: zod.number(),
  rmssd: zod.number(),
  sdnn: zod.number(),
  coherenceScore: zod.number(),
  coherenceLevel: zod.string(),
  quality: zod.string(),
  scannedAt: zod.date(),
  createdAt: zod.date(),
});
export const GetAdminMemberScansResponse = zod.array(
  GetAdminMemberScansResponseItem,
);

/**
 * @summary Get global stats (admin only)
 */
export const GetAdminStatsResponse = zod.object({
  totalMembers: zod.number(),
  totalScans: zod.number(),
  avgCoherenceScore: zod.number().nullable(),
  avgHeartRate: zod.number().nullable(),
  highCoherenceCount: zod.number(),
  mediumCoherenceCount: zod.number(),
  lowCoherenceCount: zod.number(),
});

/**
 * @summary Set or remove admin role for a member
 */
export const SetAdminRoleParams = zod.object({
  userId: zod.coerce.number(),
});

export const SetAdminRoleBody = zod.object({
  isAdmin: zod.boolean(),
});

export const SetAdminRoleResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  isAdmin: zod.boolean(),
  createdAt: zod.date(),
});