# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Peace Power API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalMembers: number;
  totalScans: number;
  /** @nullable */
  avgCoherenceScore: number | null;
  /** @nullable */
  avgHeartRate: number | null;
  highCoherenceCount: number;
  mediumCoherenceCount: number;
  lowCoherenceCount: number;
}