# lib/api-zod/src/generated/types/memberWithStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Peace Power API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanRecord } from "./scanRecord";

export interface MemberWithStats {
  id: number;
  name: string;
  email: string;
  isAdmin: boolean;
  createdAt: Date;
  scanCount: number;
  lastScan: ScanRecord | null;
  /** @nullable */
  avgCoherenceScore: number | null;
  /** @nullable */
  avgHeartRate: number | null;
}