# lib/api-zod/src/generated/types/submitScanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Peace Power API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitScanBody {
  heartRate: number;
  rmssd: number;
  sdnn: number;
  coherenceScore: number;
  coherenceLevel: string;
  quality: string;
}